+++ src/LinkMobility/ILinkMobilityClient.cs
@@ -9,17 +9,14 @@ namespace AMWD.Net.Api.LinkMobility
 	public interface ILinkMobilityClient
 	{
 		/// <summary>
-		/// Sends a text message to a list of recipients.
+		/// Performs a POST request to the LINK mobility API.
 		/// </summary>
+		/// <typeparam name="TResponse">The type of the response.</typeparam>
+		/// <typeparam name="TRequest">The type of the request.</typeparam>
+		/// <param name="requestPath">The path of the API endpoint.</param>
 		/// <param name="request">The request data.</param>
+		/// <param name="queryParams">Optional query parameters.</param>
 		/// <param name="cancellationToken">A cancellation token to propagate notification that operations should be canceled.</param>
-		Task<SendMessageResponse> SendTextMessage(SendTextMessageRequest request, CancellationToken cancellationToken = default);
-
-		/// <summary>
-		/// Sends a binary message to a list of recipients.
-		/// </summary>
-		/// <param name="request">The request data.</param>
-		/// <param name="cancellationToken">A cancellation token to propagate notification that operations should be canceled.</param>
-		Task<SendMessageResponse> SendBinaryMessage(SendBinaryMessageRequest request, CancellationToken cancellationToken = default);
+		Task<TResponse> PostAsync<TResponse, TRequest>(string requestPath, TRequest? request, IQueryParameter? queryParams = null, CancellationToken cancellationToken = default);
 	}
 }